requirements.txt
# Dependências para aplicação standalone de mensageria SIFEN

# Conexão com Oracle (thin mode - sem necessidade de Oracle Instant Client)
oracledb>=2.0.0

# RabbitMQ
pika==1.3.2

# Comunicação HTTP/SOAP
requests==2.32.3

# Processamento XML e assinatura digital
lxml
signxml==3.2.0
cryptography>=38.0.0,<39.0.0

pyOpenSSL==22.1.0

# Utilitários
python-dotenv==1.1.1
